+++ src/app/components/(movieSeatsPage)/DescriptionBox.tsx
