+++ src/app/components/(movieSeatsPage)/Seat.tsx
