+++ src/app/components/(movieSeatsPage)/SeatLegend.tsx
