+++ src/app/components/(movieSeatsPage)/SeatSelection.tsx
